# src/swagger/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "BBQueue API",
    "description": "BBQueue - очередь лучшей прожарки",
    "version": "v1"
  },
  "paths": {
...
    "/api/employee/employee": {
      "get": {
        "tags": [
          "Employee"
        ],
        "summary": "Поставляет информацию о сотруднике",
        "parameters": [
          {
            "name": "employeeId",
            "in": "query",
            "description": "",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/employee/employee_add": {
      "post": {
        "tags": [
          "Employee"
        ],
        "summary": "Добавляет нового сотрудника",
        "requestBody": {
          "description": "Структура содержащая информацию о пользователе",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeRegistryDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeRegistryDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeRegistryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/employee/role_set": {
      "post": {
        "tags": [
          "Employee"
        ],
        "summary": "Устанавливает сотруднику роль",
        "requestBody": {
          "description": "Структура содержащая идентификатор сотрудника и название роли",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeSetRoleDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeSetRoleDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeSetRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/employee/employee_to_window": {
      "post": {
        "tags": [
          "Employee"
        ],
        "summary": "Прикрепляет сотрудника к окну",
        "requestBody": {
          "description": "Структура содержит идентификатор окна для привязки сотрудника (а id сотрудника берём из JWT)",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeToWindowDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeToWindowDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeToWindowDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/window/work_state": {
      "post": {
        "tags": [
          "Window"
        ],
        "summary": "Меняет состояние окна (Открыто, Перерыв, Закрыто)",
        "requestBody": {
          "description": "На вход принимает номер окна и состояние (строковое значение)",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeWindowWorkStateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeWindowWorkStateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeWindowWorkStateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/window/add_window": {
      "post": {
        "tags": [
          "Window"
        ],
        "summary": "Создаёт новое окно",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WindowCreateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/WindowCreateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/WindowCreateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WindowCreatedIdDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/window/add_target_window": {
      "post": {
        "tags": [
          "Window"
        ],
        "summary": "Прикрепляет цель к окну",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WindowTargetCreateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/WindowTargetCreateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/WindowTargetCreateDto"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ChangeWindowWorkStateDto": {
        "required": [
          "number",
          "windowWorkState"
        ],
        "type": "object",
        "properties": {
          "number": {
            "maxLength": 6,
            "type": "string"
          },
          "windowWorkState": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "EmployeeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "externalSystemIdentity": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "active": {
            "type": "boolean"
          },
          "role": {
            "$ref": "#/components/schemas/EmployeeRole"
          }
        },
        "additionalProperties": false
      },
      "EmployeeRegistryDto": {
        "required": [
          "externalSystemId",
          "name"
        ],
        "type": "object",
        "properties": {
          "externalSystemId": {
            "maxLength": 16,
            "type": "string"
          },
          "name": {
            "maxLength": 100,
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "role": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "EmployeeRole": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "EmployeeSetRoleDto": {
        "type": "object",
        "properties": {
          "employeeId": {
            "type": "integer",
            "format": "int64"
          },
          "role": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "EmployeeToWindowDto": {
        "type": "object",
        "properties": {
          "windowId": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
      "ErrorDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "WindowCreatedIdDto": {
        "type": "object",
        "properties": {
          "windowId": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
      "WindowCreateDto": {
        "type": "object",
        "properties": {
          "number": {
            "maxLength": 6,
            "type": "string",
            "nullable": true
          },
          "description": {
            "maxLength": 256,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "WindowTargetCreateDto": {
        "type": "object",
        "properties": {
          "windowId": {
            "type": "integer",
            "format": "int64"
          },
          "targetId": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Put **_ONLY_** JWT Bearer token on textbox below!",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}